implemented add notification & get notifications for student endpoints

+++ eobrazovanje/src/main/java/com/ftn/eobrazovanje/model/PreExamDuty.java
@@ -15,9 +15,9 @@ public class PreExamDuty {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    private Integer schoolYear;
+    private int schoolYear;
 
     @ManyToOne
     private CourseTeacher courseTeacher;
